split Result component and useHooks
